MISUV-2862 fix scalastyle
