+++ app/models/LiabilityCalculation/LiabilityCalculationResponse.scala
@@ -19,7 +19,6 @@ package models.LiabilityCalculation
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
 import models.LiabilityCalculation.taxCalculation.TaxCalculation
-import models.LiabilityCalculation.reliefs.Reliefs
 import models.LiabilityCalculation.taxDeductedAtSource.TaxDeductedAtSource
 
 sealed trait LiabilityCalculationResponseModel
@@ -30,7 +29,7 @@ object LiabilityCalculationError {
 }
 
 case class LiabilityCalculationResponse(metadata: Metadata = Metadata(),
-                                        calculation: Calculation = Calculation(),
+                                        calculation: Calculation = Calculation()
                                        ) extends LiabilityCalculationResponseModel
 object LiabilityCalculationResponse {
   implicit val format: OFormat[LiabilityCalculationResponse] = Json.format[LiabilityCalculationResponse]
@@ -51,7 +50,7 @@ case class Calculation(
                         allowancesAndDeductions: AllowancesAndDeductions = AllowancesAndDeductions(),
                         reliefs: Reliefs = Reliefs(),
                         taxDeductedAtSource: TaxDeductedAtSource = TaxDeductedAtSource(),
-                          giftAid: GiftAid = GiftAid(),
+                        giftAid: GiftAid = GiftAid(),
                         marriageAllowanceTransferredIn: MarriageAllowanceTransferredIn = MarriageAllowanceTransferredIn(),
                         employmentAndPensionsIncome: EmploymentAndPensionsIncome = EmploymentAndPensionsIncome(),
                         employmentExpenses: EmploymentExpenses = EmploymentExpenses(),
@@ -62,7 +61,7 @@ case class Calculation(
                         savingsAndGainsIncome: SavingsAndGainsIncome = SavingsAndGainsIncome(),
                         dividendsIncome: DividendsIncome = DividendsIncome(),
                         incomeSummaryTotals: IncomeSummaryTotals = IncomeSummaryTotals(),
-                        taxCalculation: TaxCalculation = TaxCalculation(),
+                        taxCalculation: TaxCalculation = TaxCalculation()
                       )
 
 object Calculation {
@@ -86,55 +85,6 @@ object Calculation {
         (JsPath \ "incomeSummaryTotals").read[IncomeSummaryTotals] and
         (JsPath \ "taxCalculation").read[TaxCalculation]
       ) (Calculation.apply _)
-}
-
-
-case class AllowancesAndDeductions(
-                                    personalAllowance: Option[Int] = None,
-                                    marriageAllowanceTransferOut: MarriageAllowanceTransferOut = MarriageAllowanceTransferOut(),
-                                    reducedPersonalAllowance: Option[Int] = None,
-                                    giftOfInvestmentsAndPropertyToCharity: Option[Int] = None,
-                                    lossesAppliedToGeneralIncome: Option[Int] = None,
-                                    qualifyingLoanInterestFromInvestments: Option[BigDecimal] = None,
-                                    postCessationTradeReceipts: Option[BigDecimal] = None,
-                                    paymentsToTradeUnionsForDeathBenefits: Option[BigDecimal] = None,
-                                    grossAnnuityPayments: Option[BigDecimal] = None,
-                                    pensionContributions: Option[BigDecimal] = None
-                                  )
-object AllowancesAndDeductions {
-  implicit val writes: OWrites[AllowancesAndDeductions] = (
-    (JsPath \ "personalAllowance").writeNullable[Int] and
-      (JsPath \ "marriageAllowanceTransferOut").write[MarriageAllowanceTransferOut] and
-      (JsPath \ "reducedPersonalAllowance").writeNullable[Int] and
-      (JsPath \ "giftOfInvestmentsAndPropertyToCharity").writeNullable[Int] and
-      (JsPath \ "lossesAppliedToGeneralIncome").writeNullable[Int] and
-      (JsPath \ "qualifyingLoanInterestFromInvestments").writeNullable[BigDecimal] and
-      (JsPath \ "post-cessationTradeReceipts").writeNullable[BigDecimal] and
-      (JsPath \ "paymentsToTradeUnionsForDeathBenefits").writeNullable[BigDecimal] and
-      (JsPath \ "grossAnnuityPayments").writeNullable[BigDecimal] and
-      (JsPath \ "pensionContributions").writeNullable[BigDecimal]
-    ) (unlift(AllowancesAndDeductions.unapply))
-
-  implicit val reads: Reads[AllowancesAndDeductions] = (
-    (JsPath \ "personalAllowance").readNullable[Int] and
-      (JsPath \ "marriageAllowanceTransferOut").read[MarriageAllowanceTransferOut] and
-      (JsPath \ "reducedPersonalAllowance").readNullable[Int] and
-      (JsPath \ "giftOfInvestmentsAndPropertyToCharity").readNullable[Int] and
-      (JsPath \ "lossesAppliedToGeneralIncome").readNullable[Int] and
-      (JsPath \ "qualifyingLoanInterestFromInvestments").readNullable[BigDecimal] and
-      (JsPath \ "post-cessationTradeReceipts").readNullable[BigDecimal] and
-      (JsPath \ "paymentsToTradeUnionsForDeathBenefits").readNullable[BigDecimal] and
-      (JsPath \ "grossAnnuityPayments").readNullable[BigDecimal] and
-      (JsPath \ "pensionContributions").readNullable[BigDecimal]
-    ) (AllowancesAndDeductions.apply _)
-}
-
-case class MarriageAllowanceTransferOut(
-                                         personalAllowanceBeforeTransferOut: Option[BigDecimal] = None,
-                                         transferredOutAmount: Option[BigDecimal] = None
-                                       )
-object MarriageAllowanceTransferOut {
-  implicit val format: OFormat[MarriageAllowanceTransferOut] = Json.format[MarriageAllowanceTransferOut]
 }
 
 case class ChargeableEventGainsIncome(totalOfAllGains: Option[Int] = None)
@@ -202,8 +152,8 @@ object IncomeSummaryTotals {
 }
 
 case class MarriageAllowanceTransferredIn(
-                    amount: Option[BigDecimal] = None
-                  )
+                                           amount: Option[BigDecimal] = None
+                                         )
 object MarriageAllowanceTransferredIn {
   implicit val format: OFormat[MarriageAllowanceTransferredIn] = Json.format[MarriageAllowanceTransferredIn]
 }
@@ -222,5 +172,3 @@ case class StateBenefitsIncome(totalStateBenefitsIncome: Option[BigDecimal] = No
 object StateBenefitsIncome {
   implicit val format: OFormat[StateBenefitsIncome] = Json.format[StateBenefitsIncome]
 }
-
-
