+++ app/models/LiabilityCalculation/taxDeductedAtSource/TaxDeductedAtSource.scala
@@ -14,22 +14,6 @@
  * limitations under the License.
  */
 
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package models.LiabilityCalculation.taxDeductedAtSource
 
 import play.api.libs.json._
@@ -43,7 +27,7 @@ case class TaxDeductedAtSource(
                                 occupationalPensions: Option[BigDecimal] = None,
                                 stateBenefits: Option[BigDecimal] = None,
                                 specialWithholdingTaxOrUkTaxPaid: Option[BigDecimal] = None,
-                                inYearAdjustmentCodedInLaterTaxYear: Option[BigDecimal] = None,
+                                inYearAdjustmentCodedInLaterTaxYear: Option[BigDecimal] = None
                               )
 
 object TaxDeductedAtSource {
